+++ src/modules/products/products.service.ts
@@ -9,6 +9,7 @@ import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { Product } from './entities/product.entity';
 import { Category } from '../categories/entities/category.entity';
+import { Cart } from '../carts/entities/cart.entity';
 
 @Injectable()
 export class ProductsService {
@@ -17,9 +18,16 @@ export class ProductsService {
     private productsRepository: Repository<Product>,
     @InjectRepository(Category)
     private categoriesRepository: Repository<Category>,
+    @InjectRepository(Category)
+    private cartsRepository: Repository<Cart>,
   ) {}
 
-  async getByName(name: string, categoryId: string) {
+  /**
+   * Search for a product by name in a specific category
+   * @param name
+   * @param categoryId
+   */
+  async getByName(name: string, categoryId: string): Promise<Product> {
     return await this.productsRepository
       .createQueryBuilder('products')
       .where('products.name = :name')
@@ -28,7 +36,10 @@ export class ProductsService {
       .getOne();
   }
 
-  async getCategoryById(categoryId: string) {
+  /**
+   * @param categoryId
+   */
+  async getCategoryById(categoryId: string): Promise<Category> {
     return await this.productsRepository
       .createQueryBuilder('categories')
       .where('categories.id = :categoryId')
@@ -36,7 +47,19 @@ export class ProductsService {
       .getOne();
   }
 
-  async create(createProductDto: CreateProductDto) {
+  /**
+   * Get all products in a category
+   * @param categoryId
+   */
+  async getByCategoryId(categoryId: string): Promise<Product[]> {
+    return await this.productsRepository.find({ categoryId });
+  }
+
+  /**
+   * Add product
+   * @param createProductDto
+   */
+  async create(createProductDto: CreateProductDto): Promise<Product> {
     const product = await this.getByName(
       createProductDto.name,
       createProductDto.categoryId,
@@ -61,6 +84,9 @@ export class ProductsService {
     );
   }
 
+  /**
+   * Retrieve all products
+   */
   async findAll(): Promise<{ totalCount: number; items: Product[] }> {
     return {
       totalCount: await this.productsRepository.count(),
@@ -68,6 +94,10 @@ export class ProductsService {
     };
   }
 
+  /**
+   * Get product details
+   * @param {string} id - product ID
+   */
   async findOne(id: string): Promise<Product> {
     const product = await this.productsRepository.findOne(id);
 
@@ -78,13 +108,22 @@ export class ProductsService {
     return product;
   }
 
-  async update(id: string, updateProductDto: UpdateProductDto) {
+  /**
+   * Update product details
+   * @param id
+   * @param updateProductDto
+   */
+  async update(
+    id: string,
+    updateProductDto: UpdateProductDto,
+  ): Promise<Product> {
     const product = await this.productsRepository.findOne(id);
 
     if (!product) {
       throw new NotFoundException('Product not found.');
     }
 
+    // Check if category exists before changing category
     if (updateProductDto.categoryId) {
       const category = await this.getCategoryById(updateProductDto.categoryId);
 
@@ -100,16 +139,34 @@ export class ProductsService {
     return await this.productsRepository.save(product);
   }
 
-  async remove(id: string) {
+  /**
+   * Remove a specific product
+   * @param id
+   */
+  async remove(id: string): Promise<Product> {
     const product = await this.productsRepository.findOne(id);
 
     if (!product) {
       throw new NotFoundException('Product not found.');
     }
 
-    //todo: check product in cart
+    // Make sure product is not currently added to cart.
+    if (await this.checkIfProductIsInCart(id)) {
+      throw new NotAcceptableException('Product is currently added to cart.');
+    }
 
     await this.productsRepository.delete(id);
     return product;
+  }
+
+  /**
+   * @param productId
+   */
+  async checkIfProductIsInCart(productId: string): Promise<boolean> {
+    return !!(await this.cartsRepository
+      .createQueryBuilder('carts')
+      .where('carts.productId = :productId')
+      .setParameter('productId', productId)
+      .getCount());
   }
 }
